Added Grass Struct
 The Hidden Grasslands, C++ Intermediate
