+++ Pokemon/Game.hpp
@@ -1,0 +1,11 @@
+class Player;
+
+class Game
+{
+public:
+
+	Game();
+
+	void GameLoop(Player& player);
+		
+};
