+++ Pokemon/Pokemon.vcxproj
@@ -154,10 +154,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Game.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Player.cpp" />
     <ClCompile Include="Player.hpp" />
     <ClCompile Include="Pokemon.cpp" />
+    <ClCompile Include="ProfessorOak.cpp" />
     <ClCompile Include="Utility.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -169,7 +171,10 @@
   <ItemGroup>
     <ClInclude Include="EPokemonChoice.hpp" />
     <ClInclude Include="EPokemonType.hpp" />
+    <ClInclude Include="FGrass.hpp" />
+    <ClInclude Include="Game.hpp" />
     <ClInclude Include="Pokemon.hpp" />
+    <ClInclude Include="ProfessorOak.hpp" />
     <ClInclude Include="Utility.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
