+++ Pokemon/Pokemon/x64/Debug/Pokemon.log
@@ -1,7 +1,4 @@
-﻿  main.cpp
-  Player.cpp
-  Player.hpp
-  Pokemon.cpp
+﻿  ProfessorOak.cpp
   Utility.cpp
   Generating Code...
 Pokemon\x64\Debug\Player.obj : warning LNK4042: object specified more than once; extras ignored
